sportstore-ai/requirements.txt
fastapi
uvicorn
pydantic
pandas
scikit-learn
sqlalchemy
pymysql
python-dotenv
